optimización del código
