+++ src/Components/TablaConBuscador.tsx
