--- src/Styles/AsignarCronos.css
@@ -1,6 +1,0 @@
-.full-height-card {
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-  }
