--- src/Styles/Bootstrap.css
@@ -1,76 +1,0 @@
-@import url('https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css');
-
-/* Colores de los headers */
-.bg-unp {
-    background-color: #303d50;
-}
-
-.border-unp {
-    border-color: #365072;
-}
-
-.bg-section {
-    background-color: #f7f7f7;
-}
-
-.text-subtitle {
-    color: #303d50;
-}
-
-input[type="text"][disabled] {
-    background-color: #ffffff;
-}
-
-select:disabled {
-    background-color: #ffffff;
-}
-
-.nav-link.active {
-    background-color: #f7f7f7 !important;
-}
-
-.nav-link {
-    border-bottom: #f7f7f7 !important;
-}
-
-.btn-next {
-    color: #f3f3f3;
-    background-color: #365072;
-    border-color: #314c70
-}
-
-.btn-next:hover {
-    color: #f3f3f3;
-    background-color: #3b5f8d;
-    border-color: #314c70
-}
-
-.btn-next:active:focus {
-    color: #f3f3f3;
-    background-color: #365072;
-    border-color: #314c70
-}
-
-.btn-send {
-    color: #f3f3f3;
-    background-color: #38a729;
-    border-color: #33a024
-}
-
-.btn-send:hover {
-    color: #f3f3f3;
-    background-color: #2d8b20;
-    border-color: #2d8b21
-}
-
-.btn-send:active:focus {
-    color: #f3f3f3;
-    background-color: #38a729;
-    border-color: #33a024
-}
-
-/* Clases para sweetalert2 */
-
-.swal2-confirm {
-    background-color: #6c757d;
-}
