--- src/Styles/Bootstrap1.css
@@ -1,18 +1,0 @@
-@import url('https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css');
-
-/* Colores de los headers */
-.bg-unp {
-    background-color: #303d50;
-}
-
-.border-unp {
-    border-color: #365072;
-}
-
-.bg-section {
-    background-color: #f7f7f7;
-}
-
-.text-subtitle {
-    color: #303d50;
-}
