--- src/Styles/Fieldset1.css
@@ -1,7 +1,0 @@
-.hidden {
-  display: none;
-}
-
-.visible {
-  display: block;
-}
