--- src/Styles/Parrafos.css
@@ -1,4 +1,0 @@
-.Leyenda-autoriza-ec {
-    font-size: 0.8rem;
-    color: rgb(80, 80, 80);
-}
