--- src/Styles/Pasos.css
@@ -1,23 +1,0 @@
-.steps-container {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 20px;
-}
-
-.step {
-  width: 60px;
-  height: 60px;
-  border-radius: 50%;
-  background-color: lightgray;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 5px;
-  font-weight: bold;
-  color: white;
-  cursor: pointer;
-}
-
-.step.active {
-  background-color: #d36269;
-}
