--- src/Styles/PasosIconos.css
@@ -1,78 +1,0 @@
-.card {
-  z-index: 0;
-}
-
-#progressbar {
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  margin-bottom: 25px;
-  margin-top: 10px;
-  padding: 0;
-  list-style-type: none;
-  color: lightgrey;
-}
-
-#progressbar .active {
-  color: #d36269;
-}
-
-#progressbar li {
-  font-size: 12px;
-  width: 10%;
-  position: relative;
-  text-align: center;
-  font-weight: 400;
-}
-
-#progressbar li .step {
-  font-size: 15px;
-  color: gray;
-  width: 60px;
-  height: 60px;
-  border-radius: 50%;
-  background-color: lightgray;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto 10px auto;
-  font-weight: normal;
-  color: white;
-  cursor: pointer;
-}
-
-/*Icon ProgressBar before any progress*/
-#progressbar li:before {
-  width: 50px;
-  height: 50px;
-  line-height: 45px;
-  display: block;
-  font-size: 12px;
-  color: #ffffff;
-  background: lightgray;
-  border-radius: 50%;
-  margin: 0 auto 10px auto;
-  padding: 2px;
-}
-
-/*ProgressBar connectors*/
-#progressbar li:after {
-  content: '';
-  width: 100%;
-  height: 3px;
-  background: lightgray;
-  position: absolute;
-  left: 0;
-  top: 30px;
-  z-index: -1;
-  font-size: 12px;
-}
-
-#progressbar li.active:before, #progressbar li.active:after {
-  background: #d36269;
-}
-
-#progressbar li.active .step {
-  background: #d36269;
-  color: white;
-}
