--- src/Styles/Slider.css
@@ -1,25 +1,0 @@
-.slide-enter {
-  max-height: 0;
-  opacity: 0;
-  overflow: hidden;
-  transition: max-height 300ms ease-out, opacity 300ms ease-out;
-}
-
-.slide-enter-active {
-  max-height: 500px;
-  /* Ajusta según sea necesario */
-  opacity: 1;
-}
-
-.slide-exit {
-  max-height: 500px;
-  /* Ajusta según sea necesario */
-  opacity: 1;
-  overflow: hidden;
-  transition: max-height 300ms ease-in, opacity 300ms ease-in;
-}
-
-.slide-exit-active {
-  max-height: 0;
-  opacity: 0;
-}
