--- src/Styles/Tabla.css
@@ -1,156 +1,0 @@
-:root {
-    --color-dia-unodos: #67b459;
-    --color-dia-tres: #EEE910;
-    --color-dia-cuatro: #F69300;
-    --color-dia-cinco: #EB2D32;
-}
-
-th .d-th {
-    min-width: 250px;
-}
-
-th .c-th {
-    min-width: 80px;
-}
-
-.th-solicitud {
-    padding-left: 10px;
-}
-
-tr:nth-child(odd) td {
-    background-color: #F7F7F7;
-    border: 1px solid #FFFFFF;
-    border-left: 1px solid #F7F7F7;
-    border-right: 1px solid #F7F7F7;
-}
-
-tr:nth-child(even) td {
-    border: 1px solid #FFFFFF;
-}
-
-.icono {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 50%;
-    background-color: #303d50cc;
-    border: 1px solid #ffffff;
-    width: 40px;
-    height: 40px;
-    margin-right: 10px;
-}
-
-.icono .fa-icono {
-    width: 22px;
-    height: 22px;
-    color: #ffffff;
-}
-
-.r-tabla {
-    display: flex;
-    align-items: center;
-    min-height: 70px;
-    max-height: 70px;
-}
-
-.r-tabla-solicitud {
-    padding-left: 5px;
-}
-
-.td-solicitud {
-    text-align: center;
-    vertical-align: middle;
-    min-width: 100px;
-    color: #303D50;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-solicitud:hover {
-    background-color: #029a465e;
-}
-
-.r-tabla .solicitud .numero-solicitud {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.td-radicado {
-    min-width: 170px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-radicado:hover {
-    background-color: #029a465e;
-}
-
-.td-dhabiles {
-    min-width: 150px;
-}
-
-.td-nombre {
-    min-width: 180px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-nombre:hover {
-    background-color: #029a465e;
-}
-
-.td-identificacion {
-    min-width: 200px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-identificacion:hover {
-    background-color: #029a465e;
-}
-
-.td-genero {
-    min-width: 150px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-genero:hover {
-    background-color: #029a465e;
-}
-
-.td-contacto,
-.td-ubicacion {
-    min-width: 250px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-direccion {
-    min-width: 350px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.td-contacto:hover,
-.td-ubicacion:hover,
-.td-direccion:hover {
-    background-color: #029a465e;
-}
-
-.r-tabla .radicado {
-    font-weight: 400;
-}
-
-.r-tabla .tiempo-estado {
-    width: 90px;
-    text-align: center;
-    color: #ffffff;
-    font-weight: 500;
-    border-radius: 40px;
-    padding: 6px;
-    border: 1px solid #ffffff;
-    background-color: var(--color-dia-unodos);
-}
