--- src/Styles/TabsLaterales.css
@@ -1,77 +1,0 @@
-.tabs-laterales {
-    position: fixed;
-    top: 0;
-    right: 0;
-    height: 100%;
-    display: flex;
-    align-items: center;
-  }
-  
-  .tabs {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-end;
-    position: relative;
-    z-index: 1;
-  }
-  
-  .tab-btn {
-    background-color: #007bff;
-    color: white;
-    padding: 10px 15px;
-    margin: 5px 0;
-    border: none;
-    cursor: pointer;
-    transform: rotate(-90deg);
-    transform-origin: right;
-    position: relative;
-    right: -60px;
-    font-size: 16px;
-    z-index: 2;
-  }
-  
-  .tab-btn:hover {
-    background-color: #0056b3;
-  }
-  
-  .tab-content {
-    position: fixed;
-    top: 0;
-    right: -300px; /* Fuera de la pantalla por defecto */
-    width: 300px;
-    height: 100%;
-    background-color: #f4f4f4;
-    box-shadow: -2px 0 5px rgba(0, 0, 0, 0.5);
-    padding: 20px;
-    transition: right 0.3s ease; /* Animación de deslizamiento */
-    z-index: 1;
-  }
-  
-  .tab-content.active {
-    right: 0; /* Aparece cuando la pestaña está activa */
-  }
-  
-  form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  form input {
-    margin-bottom: 10px;
-    padding: 8px;
-    font-size: 16px;
-  }
-  
-  form button {
-    padding: 10px;
-    background-color: #28a745;
-    color: white;
-    border: none;
-    cursor: pointer;
-  }
-  
-  form button:hover {
-    background-color: #218838;
-  }
-
