--- src/Styles/prueba.css
@@ -1,1 +1,0 @@
-/* //sssdf */
