--- src/Styles/ventanaLateral.css
@@ -1,42 +1,0 @@
-.ventana-lateral {
-    display: flex;
-    align-items: start;
-    justify-content: center;
-    border-radius: 7px;
-    /* background-color: #F7F7F7; */
-    min-height: calc(100vh - 138px);
-    max-height: calc(100vh - 138px);
-    overflow-y: auto;
-}
-
-.ventana-tabla {
-    display: flex;
-    align-items: start;
-    justify-content: flex-start;
-    border-radius: 7px;
-    min-height: calc(100vh - 300px);
-    max-height: calc(100vh - 138px);
-    padding: 0px;
-    overflow-y: auto;
-    overflow-x: auto;
-    scrollbar-width: none;
-    -ms-overflow-style: none;
-}
-
-
-@supports (-ms-ime-align: auto) {
-    .ventana-tabla {
-        scrollbar-width: none; /* Oculta la barra de desplazamiento en Edge */
-        -ms-overflow-style: none; /* Oculta la barra de desplazamiento en Edge */
-    }
-}
-
-.leaflet-container {
-    height: calc(100vh - 140px);
-    width: 100%;
-    border-radius: 7px;
-}
-
-.leaflet-control-attribution {
-    display: none;
-}
